Change to original repository for api/monitoring
Version v0.44.1 of prometheus-operator didn't have a go.mod file for
pkg/client, which is likely why it is still included with this version
number. For this reason reverting the change to use the ZipRecruiter
module and use the replace line to pick up the current code. Hopefully
this satisfies updating modules.